Added very basic remote handling
Added a new SiriRemoteView which has support for the basic swipe gesture handlers.  This is more of a functional test.  It works great, but from an api standpoint it's a bit bloated.  I could probably reduce the onSwipe[Left|Right|Up|Down] to a single onSwipe event which contains direction.

I debated handling presses in a similar fashion to the RCTTouchHandler but I don't think that model works for the tvOS environment.  Next up is handling raw presses.

PS.  This works great with the Game2048 demo.

Signed-off-by: Timothy Braun <braunsquared@gmail.com>